completed day 15 HR question and also fixed Project error
Project error was when we were getting at the top of the file : "The class/package
 was not a project file and hence only syntax error were being highlighted"

Solution : Was to have no workspace with multiple java files. Instead of organizing all the java files into respective folders, I simply deleted the workspace itself.
Just open the folder every time you need to work on the files of that folder.

Ez pz